+++ nikitov.dmitriy/S1/output_list.cpp
@@ -2,9 +2,9 @@
 #include <fstream>
 #include "iterator.hpp"
 
-void nikitov::outputList(List< std::pair< std::string, List< int > > >& pairsList, std::ostream& output)
+void nikitov::outputList(List< std::pair< std::string, List< size_t > > >& pairsList, std::ostream& output)
 {
-  ListIterator< std::pair< std::string, List< int > > > pairsIterator = pairsList.begin();
+  ListIterator< std::pair< std::string, List< size_t > > > pairsIterator = pairsList.begin();
   size_t maxSize = 0;
   for (size_t i = 0; i != pairsList.size(); ++i)
   {
@@ -19,21 +19,21 @@ void nikitov::outputList(List< std::pair< std::string, List< int > > >& pairsLis
   output << '\n';
 
   size_t sizeOfPairs = pairsList.size();
-  int* sums = new int[sizeOfPairs]{};
+  size_t* sums = new size_t[sizeOfPairs]{};
   for (size_t i = 0; i != maxSize; ++i)
   {
     int sum = 0;
     pairsIterator = pairsList.begin();
     for (size_t j = 0; j != sizeOfPairs; ++j)
     {
-      ListIterator< int > numbersIterator = pairsIterator->second.begin();
+      ListIterator< size_t > numbersIterator = pairsIterator->second.begin();
       if (i < pairsIterator->second.size())
       {
         if (sum != 0)
         {
           output << ' ';
         }
-        int number = *(numbersIterator.advance(i));
+        size_t number = *(numbersIterator.advance(i));
         sum += number;
         output << number;
       }
